<리펙토링> (Step1) : 문자열 덧셈 계산기
구분자 관리 / 구분자를 통한 문자 분리 / 유효성 체크 기준으로 리펙토링
