+++ src/main/kotlin/DelimiterManager.kt
@@ -6,10 +6,8 @@ class DelimiterManager {
         val matchResult = regex.find(text)
 
         return if (matchResult != null) {
-            val delimiter = matchResult.groupValues[1]
-            delimiters.add(delimiter)
-            val remainingText = matchResult.groupValues[2]
-            remainingText
+            delimiters.add(matchResult.groupValues[1])
+            matchResult.groupValues[2]
         } else {
             text
         }
