--- src/main/kotlin/FormulaValidator.kt
@@ -1,5 +1,0 @@
-class FormulaValidator {
-    fun isNullOrEmptyOrBlank(text: String?): Boolean {
-        return text.isNullOrEmpty() || text.isBlank()
-    }
-}
